Done
There is a nine floor building with 4 entrances. The entrance number starts with one. There are 4 apartments on one floor. Write a program that will receive the apartment number from the keyboard, and display on the screen on which floor, which entrance this apartment is located. If there is no such apartment in this house, inform the user.